feat: support decimal quantities for yard/meter unit products
Qty inputs across add/edit product, restock, quick-stock, POS cart
editing, purchase orders, returns, and CSV import all truncated to
whole numbers via parseInt regardless of the product's unit, so a
2.5-yard batch created via the POS variant picker (already
decimal-aware) would get floored the moment it was touched anywhere
else. Thread `unit`/isDecimal checks through each of those inputs and
switch to parseFloat + 0.1 step when the unit is yard or meter.

Also widen purchase_order_items.quantity from integer to numeric —
it was the only stock-quantity column not already numeric, so
fractional-unit POs would have been rejected at the DB level even
with the UI fixed.

+++ src/components/products/ProductImporter.tsx
@@ -196,7 +196,7 @@ export function ProductImporter({ onClose, onSuccess }: ProductImporterProps) {
                     }
 
                     // Create or Update Batch
-                    const qty = parseInt(row.quantity || '0');
+                    const qty = parseFloat(row.quantity || '0');
                     const costPrice = parseFloat(row.cost_price);
                     const markup = parseFloat(row.markup_percentage);
                     const sellingPrice = costPrice * (1 + markup / 100);
